+++ app/api/search/route.ts
@@ -1,11 +1,12 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { Prisma } from '@prisma/client';
 import prisma from '~/lib/prisma';
+import { authorInclude, recipeSelect } from '~/lib/prisma/index';
 import { searchSchema } from '~/utils/validation/search';
 import { STATUS_CODE } from '~/constants/api';
 import { SearchQueryKey, SearchTabKey, SearchTabValue } from '~/constants/key';
 import { ErrorResponse } from '../lib/common';
-import { Author, authorInclude, Recipe, recipeSelect } from '../main/route';
+import { Author, Recipe } from '../main/route';
 
 export type Ingredient = Prisma.RecipeGetPayload<object>;
