+++ lib/prisma/index.ts
@@ -1,0 +1,25 @@
+export const recipeSelect = {
+  id: true,
+  title: true,
+  thumbnailUrl: true,
+  tags: true,
+  viewCount: true,
+  author: {
+    select: {
+      name: true,
+    },
+  },
+  _count: {
+    select: {
+      likes: true,
+    },
+  },
+} as const;
+
+export const authorInclude = {
+  _count: {
+    select: {
+      Recipe: true,
+    },
+  },
+} as const;
